pyproject.toml
[project]
name = "autotext"
version = "0.1.0"
description = "AutoML Pipeline for Text Classification"
requires-python = ">=3.10"

dependencies = [
    # Core ML and DL frameworks
    "torch>=2.0.0",
    "transformers>=4.30.0",
    
    # HPO and optimization
    "optuna>=3.2.0",
    
    # Data processing
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.3.0",
    
    # Configuration and utilities
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    
    # Text preprocessing
    "nltk>=3.8",
]